Simplify GPU RD PARAMETERS structure
When GPU is enabled, mv cost tables, Switchable interp
filters cost tables, mode cost table, ... are signaled
for use in fullpel evaluation and rd evaluations. These
cost tables are adapted basing on counts information that
are collected at the end of frame encoding. But, in
asynchronous design where GPU processes fullpel for a
future frame it may not be possible to receive updated
cost tables. For simplicity we maintain these tables
the same till the end of sequence. In such a case, there
is no need to pass this info. over and over again. This
changes cleans up that bit and is expected to simplify
GPU cache map overheads.

Change-Id: Ie1b16b68ccd46e91e6c4c1988f417928f856913b