Enqueue GPU jobs from a separate thread
In earlier case, the master thread or one of the worker
thread enqueues the Jobs to GPU. As enqueuing process can
be handled independently, this process is forked on a new
thread. This process although being very light, if the OS
schedules this on other active core then encoding threads
will be effectively focus on frame encoding.

Change-Id: I4d42fb39ae4b182db204968f0f8ce8f490239213