simplify prediction of cyclic refresh blocks
When GPU is enabled, the newmv analysis of future frame
happens in GPU while the current frame is being encoded.
The newmv analysis and choose partitioning modules require
segment information for their tasks. But, the segment
information of future frame is computed only after completely
coded. So we estimate the refresh map of future frame in advance.
The estimation logic is simplified. This affects quality.
We can revert it at a later stage if necessary.

Change-Id: I7405d720b51d6e2ac765008cd84dc39682cfbb4c